% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticPower.R
\name{LogisticLogLikeHood}
\alias{LogisticLogLikeHood}
\title{Log Likelihood Function for Logistic Model}
\usage{
LogisticLogLikeHood(Nd0, Nd1, Nd2, Nh0, Nh1, Nh2, alpha = 0.5, beta = 0.01)
}
\arguments{
\item{Nd0}{Number of disease subjects with 0 risk alleles}

\item{Nd1}{Number of disease subjects with 1 risk allele}

\item{Nd2}{Number of disease subjects with 2 risk alleles}

\item{Nh0}{Number of healthy subjects with 0 risk alleles}

\item{Nh1}{Number of healthy subjects with 1 risk allele}

\item{Nh2}{Number of healthy subjects with 2 risk alleles}

\item{alpha}{Intercept}

\item{beta}{Odds Ratio for Genotype Effect}
}
\description{
This function returns the negative log likelihood for logistic regression
}
\examples{
LogisticLogLikeHood()
}
