% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeN.R
\name{N.QuantitativeTrait.Numeric}
\alias{N.QuantitativeTrait.Numeric}
\title{Sampled Size Required for Powering a Genetic Association for a Quantiative Trait with Numeric Genotype Count}
\usage{
N.QuantitativeTrait.Numeric(
  power = 0.8,
  delta = 1,
  freq = 0.15,
  sigma = 1,
  NCovariates = 0,
  alpha = 0.05,
  numtests = 1,
  minh = "additive",
  moi = 0.5,
  rsquared = NULL
)
}
\arguments{
\item{power}{the required statistical power as a decimal (e.g., for 80\% power, use 0.8)}

\item{delta}{Effect size difference between homozygote risk/disease allele vs. homozygote reference allele, aka: mean(bb) - mean (AA), where 'b' is the disease allele, 'A' is the reference allele}

\item{freq}{allele frequency of disease (alternate) allele 'b'}

\item{sigma}{standard deviation of the response phenotype}

\item{NCovariates}{the number of additional parameters in the model (will reduce your overall degrees of freedom)}

\item{alpha}{desired significance level of association. Default is 0.05}

\item{numtests}{the number of tests to be corrected by Bonferroni adjustment before achieving 'alpha'}

\item{minh}{mode of inheritance:  "recessive", "additive", "dominant" (this is the same as moi=0,0.5, and 1.0, respectively). Default is "additive" if no moi is specified.}

\item{moi}{mode of inheritance: 0 for recessive, 0.5 for additive, 1.0 for dominant, or anywhere in between. This parameter OVER-RIDES minh, and is useful for modeling in-between moi's.}

\item{rsquared}{fraction of total sum-of-squares explained by model fit - OVER-RIDES delta AND sigma. Useful if the only benchmark summary statistics available is an R2 value.}
}
\description{
This function returns the required sample size to power a linear regression of genetic association for a quantitative trait and a numeric treatment of genotype count.
}
\examples{
N.QuantitativeTrait.Numeric()
}
